Working on 790
Added description attribute to query. 
I.e. when a query is created on the queries section to add to the previous queries section, it will now also have a description attribute set to "test".
In the future, this description attribute can be a blank string, which could be updated with the inputted description on the previous queries page.